Tue, Apr 19, 2022, 6:57 AM +00:00

+++ sentimental-cash/cash.py
@@ -7,9 +7,9 @@ while True:
         break
 
 change = change * 100
-qu = change / 25
-di = (change - (qu * 25)) / 10
-ni = (change - (qu * 25) - (di * 10)) / 5
-pe = (change - (qu * 25) - (di * 10) - (ni * 5))
+qu = int(change / 25)
+di = int((change - (qu * 25)) / 10)
+ni = int((change - (qu * 25) - (di * 10)) / 5)
+pe = int((change - (qu * 25) - (di * 10) - (ni * 5)))
 coins = qu + di + ni + pe
 print(coins)
